Moved all the device actuation in transactive base class into separate class
